fix: Make the hub's shutdown release the drain, and typecheck every GOOS
Three CI jobs failed on main after #342, each for its own reason. Fixing
them turned up the same shape underneath two of them: a guarantee that
held for the case someone tested and not for the case next to it.

**Hub shutdown hung on the worker's Connect stream (macos, `tunnel`).**
Cancelling the worker's connection context does not end its bidi stream:
connect-go checks the context before each read, so a Receive already
parked inside the HTTP/2 response body has to be woken by the transport
-- and once the response headers have arrived (the Hub greets every
worker with its identity) no per-stream context watcher is left to do
it. The Hub's Connect handler then saw nothing until its 10s idle
timeout, and because an unencrypted-HTTP/2 connection counts as ACTIVE
for its whole life, `http.Server.Shutdown` waited on the handler holding
it and returned its deadline as an error. #342 only changed when the
race was lost: the test now reaches `Stop()` milliseconds after the
worker connects, so the Hub reaches its drain first.

- `Client.closeStreamOnCancel` closes both stream halves when the
  connection context fires, turning every cancellation -- reconnect,
  identity watchdog, worker shutdown -- into a stream end the Hub sees
  at once. Without it the identity watchdog's "force a reconnect" could
  park forever, since Connect never returned for the reconnect loop to
  run. A send in flight is truncated rather than framed whole; both
  shapes surface to the Hub as a receive error, which is what releases
  its handler, and flushing first would re-park on the stall being
  escaped.
- `NotifyShutdownAndFence` (renamed: it fences, and the name owes the
  reader that) ends the connections it notified, on the deadline path
  too, so the Hub's exit depends on the Hub alone rather than on every
  worker choosing to close.
- Fencing latches the registry closed, and `Register` checks that latch
  inside the same critical section that publishes. A handler that
  cleared the shutdown interceptor -- a one-shot check -- a moment
  before, then spent a store round trip getting to `Register`, used to
  publish a connection nothing would ever fence, and held the drain to
  the full idle timeout. Either `Register` sees the latch or the fence
  sees the connection; there is no third outcome. The refusal is
  `Unavailable`, and comes before the greeting write.
- `Unregister` now waits for an in-flight send on both paths. It skipped
  the wait when a reconnect had already replaced the connection, which
  is precisely when a wedged send is likely -- and returning from the
  handler with a send still inside `Stream.Send` hands net/http a write
  against a finished handler, whose response state has by then been
  recycled into a pool.
- `WriteByteTimeout` bounds a peer that stops draining its socket. One
  wedged worker otherwise holds the connection mutex for the process's
  life, blocking every other sender to it. A peer that stalls its
  flow-control window instead is not bounded by this; that residual and
  its fix are tracked in #344.

**Windows build (`shell_test.go` typecheck).** #342 put `maxOnlyModelID`
/ `effortTestCatalog` in `claude_test.go`, which is `//go:build unix`,
and used them from the untagged `shell_test.go`.

- The effort fixtures and all six effort-resolver tests move to
  `claude_effort_test.go`, and the platform-neutral catalog and
  option-group tests to `claude_catalog_test.go`, both untagged. Only
  the tests that shell out to a fake CLI stay behind the tag. The
  Windows test binary gains 45 test functions (530 -> 575).
- `task lint` now runs `go vet` for windows, linux and darwin. Nothing
  else typechecked a non-host GOOS, so this class of break was
  discoverable only from a CI job the author cannot reproduce locally.
  `go build` does not catch it; `go vet` does, because it compiles
  tests.

**`TestListTerminals_AltScreenRecoveryAfterRingWrap` (linux/arm64).**
The test asserted an exact `screen_end_offset` while a real interactive
login shell wrote its own prompt into the same PTY (dash's "$ " is the 2
bytes CI saw). It was really asserting that the prompt had not arrived
yet -- true on an idle box, routine to lose under `t.Parallel`.

- `testutil.FreezeTerminalOutput` stops the PTY, waits for the shell to
  be reaped and its reader to drain, and reads the baseline back, so the
  expected offset is exact rather than assumed. The package had a second
  copy of this dance that omitted the exit wait, which is load-bearing:
  the exit notice is appended through the same counter.
- Freezing means neither alt-screen test covered a live terminal, which
  is the case the feature exists for -- a user with vim open hitting
  refresh. A live-terminal test now covers the prefix, asserting no
  offset so the prompt cannot race it.

**Environment pinning.** `NewGitCmd` layered `GIT_OPTIONAL_LOCKS=0` over
an inherited value, so a worker launched from a LeapMux terminal handed
git two entries. `exec.Cmd` resolves duplicates last-wins, so the child
was fine and only the assertion was not -- which is why this surfaced as
a test failing on a developer's own machine.

- `envutil.PinEnv` derives the keys to filter from the assignments it is
  given, so a pin cannot be half-updated. All three spawn sites the
  gitutil comment names now use it; two of them were still layering, and
  the terminal one shipped a duplicate `TERM` on essentially every
  launch.
- `FinalizeAgentEnv`'s "exactly once" test fed a fixture with nothing to
  displace, so it could not observe the duplicate production emitted.
- `envutil.ValuesFor` replaces three hand-rolled copies of "collect the
  values for this key", one of which matched case-sensitively against a
  filter that folds.

**Comments corrected.** net/http does not hijack an unencrypted-HTTP/2
connection -- it serves it inline and leaves it in `activeConn`. A
hijacked *websocket* is what escapes `Close()`, which is why
`CloseAccepted` exists. Three copies of the old claim in `locallisten`
said the opposite, and a maintainer trusting them would have removed the
call that matters. Also: the solo stop budget's "tens of milliseconds"
was off by 20x (a healthy stop is ~1.1s, nearly all of it net/http's
fixed 1s GOAWAY grace), so the budget is now stated in measured terms
and widened to 8s.

**OAuth callback.** Found while auditing what else can outlive a
shutdown: the token exchange runs on `http.DefaultClient`, which has no
timeout, on a mux route that bypasses the interceptor chain. A hung
identity provider parks a request net/http counts as ACTIVE and burns
the whole drain budget -- the same failure, one handler over. It is
bounded by `APITimeout` now; bounding every handler structurally is
tracked in #343.
